Bring dhcp up to date, new configuration scheme.
No need to use a dhcpd.conf style file anymore, dhcp_config.ml should be much
more flexible while not being so much harder to configure.

This saves me from updating the parser for all the "200 and something" dhcp
options and removes the KV_RO requirement from dhcp.

This is a compilation of the changes from haesbaert/mirage-skeleton:master, since it
diverged a little, in the future I'll work on mirage/mirage-skeleton.