docteur: use local git repository
This commit makes the docteur example use the local git repository. The
github repository may be unavailable due to github being down, no (ipv4)
internet connection or if the dns resolver only supports udp.

Due to limitations in dune we cannot use the .git/ directory directly. See
https://github.com/mirage/mirage/issues/1445 for more details.
Furthermore, git will not accept a file named .git, so a symlink to the
root .git/ directory is created with the name dotgit.